fix: remove string unions from component spec
